requirements.txt
django == 2.1.1
djangorestframework == 3.8.2
requests
